后台管理系统完成95%

+++ backend/src/main/java/com/freedy/backend/controller/CategoryController.java
@@ -3,9 +3,13 @@ package com.freedy.backend.controller;
 import java.util.Arrays;
 import java.util.Map;
 
+import com.freedy.backend.common.utils.AuthorityUtils;
 import com.freedy.backend.common.utils.Local;
 import com.freedy.backend.common.utils.Result;
+import com.freedy.backend.exception.NoPermissionsException;
+import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import com.freedy.backend.entity.CategoryEntity;
@@ -26,9 +30,8 @@ public class CategoryController {
     @Autowired
     private CategoryService categoryService;
 
-    /**
-     * 列表
-     */
+
+    @ApiOperation("列出所有分类")
     @GetMapping("/list")
     public Result list(@RequestParam Map<String, Object> params){
         PageUtils page = categoryService.queryPage(params);
@@ -36,20 +39,7 @@ public class CategoryController {
         return Result.ok().put("page", page);
     }
 
-
-    /**
-     * 信息
-     */
-    @GetMapping("/info/{id}")
-    public Result info(@PathVariable("id") Integer id){
-		CategoryEntity category = categoryService.getById(id);
-
-        return Result.ok().put("category", category);
-    }
-
-    /**
-     * 保存
-     */
+    @ApiOperation("保存分类")
     @PostMapping("/save")
     public Result save(@RequestBody CategoryEntity category){
         category.setCreatorId(Local.MANAGER_LOCAL.get().getId());
@@ -57,22 +47,33 @@ public class CategoryController {
         return Result.ok();
     }
 
-    /**
-     * 修改
-     */
+    @ApiOperation("更新分类")
     @PostMapping("/update")
     public Result update(@RequestBody CategoryEntity category){
+        //修改他人且没权限
+        if (!category.getCreatorId().equals(Local.MANAGER_LOCAL.get().getId())&&
+                !AuthorityUtils.hasAuthority("tag-operation-to-others"))
+            throw new NoPermissionsException();
 		categoryService.updateById(category);
 
         return Result.ok();
     }
 
-    /**
-     * 删除
-     */
+    @ApiOperation("删除分类")
     @GetMapping("/delete")
     public Result delete(Integer[] ids){
-		categoryService.removeByIds(Arrays.asList(ids));
+        boolean exception=false;
+        for (Integer id : ids) {
+            if (!id.equals(Local.MANAGER_LOCAL.get().getId())&&
+                    !AuthorityUtils.hasAuthority("tag-operation-to-others")){
+                exception=true;
+            }else {
+                categoryService.removeByIds(Arrays.asList(ids));
+            }
+        }
+        if (exception){
+            throw new NoPermissionsException();
+        }
         return Result.ok();
     }
